rename repositories file names
