+++ src/repositories/UserRepository/user.repository.ts
@@ -1,6 +1,6 @@
 import type { IHttpClient } from '../../services/HttpClient/IHttpClient';
-import type { IUsersRepository } from './users.interface';
-import type { User } from './users.models';
+import type { IUsersRepository } from './user.interface';
+import type { User } from './user.models';
 
 export class UsersRepository implements IUsersRepository {
   constructor(private httpClient: IHttpClient) {}
